DL-2663 - Updated Dependencies + SBT update
