+++ build.sbt
@@ -1,0 +1,45 @@
+import scoverage.ScoverageKeys
+import uk.gov.hmrc.DefaultBuildSettings.targetJvm
+import uk.gov.hmrc.sbtdistributables.SbtDistributablesPlugin.publishingSettings
+import uk.gov.hmrc.versioning.SbtGitVersioning.autoImport.majorVersion
+
+lazy val appName = "cc-calculator"
+
+lazy val appDependencies : Seq[ModuleID] = ???
+lazy val plugins : Seq[Plugins] = Seq(
+  SbtAutoBuildPlugin,
+  SbtGitVersioning,
+  SbtDistributablesPlugin,
+  SbtArtifactory
+)
+lazy val playSettings : Seq[Setting[_]] = Seq.empty
+
+lazy val scoverageSettings = {
+  Seq(
+    ScoverageKeys.coverageExcludedPackages :=  "<empty>;Reverse.*;.*Routes.*;routes_routing.*;uk.gov.hmrc;config.*;",
+    ScoverageKeys.coverageMinimum := 95,
+    ScoverageKeys.coverageFailOnMinimum := true,
+    ScoverageKeys.coverageHighlighting := true,
+    parallelExecution in Test := false
+  )
+}
+
+lazy val microservice = Project(appName, file("."))
+  .enablePlugins(Seq(play.sbt.PlayScala) ++ plugins : _*)
+  .settings(playSettings ++ scoverageSettings : _*)
+  .settings(PlayKeys.playDefaultPort := 9372)
+  .settings(majorVersion := 1)
+  .settings(publishingSettings : _*)
+  .settings(
+    targetJvm := "jvm-1.8",
+    scalaVersion := "2.12.10",
+    libraryDependencies ++= AppDependencies(),
+    retrieveManaged := true,
+    evictionWarningOptions in update := EvictionWarningOptions.default.withWarnScalaVersionEviction(false),
+    routesGenerator := InjectedRoutesGenerator,
+    resolvers := Seq(
+      Resolver.bintrayRepo("hmrc", "releases"),
+      Resolver.typesafeRepo("releases"),
+      Resolver.jcenterRepo
+    )
+  )
