+++ conf/application.conf
@@ -1,4 +1,4 @@
-# Copyright 2019 HM Revenue & Customs
+# Copyright 2020 HM Revenue & Customs
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -51,7 +51,7 @@ play.http.filters = "uk.gov.hmrc.play.bootstrap.filters.MicroserviceFilters"
 # ~~~~~
 # The secret key is used to secure cryptographics functions.
 # If you deploy your application to several instances be sure to use the same key!
-application.secret = "FlCsOdSGQReU6wlTPp3yRGrd1qTVWIx89JvBzil1JCnwq9DgknO1vXqI3eUqit5e"
+play.http.secret.key = "FlCsOdSGQReU6wlTPp3yRGrd1qTVWIx89JvBzil1JCnwq9DgknO1vXqI3eUqit5e"
 
 # Session configuration
 # ~~~~~
