--- project/MicroService.scala
@@ -1,57 +1,0 @@
-import play.routes.compiler.InjectedRoutesGenerator
-import play.sbt.routes.RoutesKeys.routesGenerator
-import sbt.Keys._
-import sbt._
-import scoverage._
-import uk.gov.hmrc.{SbtArtifactory, SbtAutoBuildPlugin}
-import uk.gov.hmrc.sbtdistributables.SbtDistributablesPlugin
-import uk.gov.hmrc.sbtdistributables.SbtDistributablesPlugin.publishingSettings
-import uk.gov.hmrc.versioning.SbtGitVersioning
-import uk.gov.hmrc.versioning.SbtGitVersioning.autoImport.majorVersion
-
-trait MicroService {
-
-  val appName: String
-
-  lazy val appDependencies : Seq[ModuleID] = ???
-  lazy val plugins : Seq[Plugins] = Seq(
-    SbtAutoBuildPlugin,
-    SbtGitVersioning,
-    SbtDistributablesPlugin,
-    SbtArtifactory
-  )
-  lazy val playSettings : Seq[Setting[_]] = Seq.empty
-
-  lazy val scoverageSettings = {
-    Seq(
-      ScoverageKeys.coverageExcludedPackages :=  "<empty>;Reverse.*;.*Routes.*;routes_routing.*;uk.gov.hmrc;config.*;",
-      ScoverageKeys.coverageMinimum := 95,
-      ScoverageKeys.coverageFailOnMinimum := true,
-      ScoverageKeys.coverageHighlighting := true,
-      parallelExecution in Test := false
-    )
-  }
-  import uk.gov.hmrc.versioning.SbtGitVersioning.autoImport.majorVersion
-
-
-  lazy val microservice = Project(appName, file("."))
-    .enablePlugins(Seq(play.sbt.PlayScala) ++ plugins : _*)
-    .settings(playSettings ++ scoverageSettings : _*)
-    .settings(majorVersion := 1)
-    .settings(publishingSettings : _*)
-    .settings(
-      libraryDependencies ++= appDependencies,
-      retrieveManaged := true,
-      evictionWarningOptions in update := EvictionWarningOptions.default.withWarnScalaVersionEviction(false),
-      routesGenerator := InjectedRoutesGenerator,
-      resolvers := Seq(
-        Resolver.bintrayRepo("hmrc", "releases"),
-        Resolver.typesafeRepo("releases"),
-        Resolver.jcenterRepo
-      ),
-      scalaVersion := "2.11.12",
-      crossScalaVersions := Seq("2.11.11"),
-      ivyScala := ivyScala.value map { _.copy(overrideScalaVersion = true) }
-    )
-
-}
